chapter2/src/main/java/org/cong/chapter2/service/CustomerService.java
package org.cong.chapter2.service;

import java.sql.*;
import java.util.*;

import org.cong.chapter2.helper.DatabaseHelper;
import org.cong.chapter2.model.Customer;
import org.cong.chapter2.util.PropsUtil;
import org.slf4j.Logger;
import org.slf4j.LoggerFactory;

/**
 * Created by liaojuncong on 4/4/16.
 */
public class CustomerService {

    private static final Logger LOGGER = LoggerFactory.getLogger(CustomerService.class);

    public List<Customer> getCustomerList() {
        String sql = "SELECT * FROM customer";
        return DatabaseHelper.queryEntityList(Customer.class, sql);
    }

    public Customer getCustomer(long id) {
        String sql = "SELECT * FROM customer WHERE id=?";
        return DatabaseHelper.queryEntity(Customer.class, sql, id);
    }

    public boolean createCustomer(Map<String, Object> fiedMap) {
        return DatabaseHelper.insertEntity(Customer.class, fiedMap);
    }

    public boolean updateCustomer(long id, Map<String, Object> fieldMap) {
        return DatabaseHelper.updateEntity(Customer.class, id, fieldMap);
    }

    public boolean deleteCustomer(long id) {
        return DatabaseHelper.deleteEntity(Customer.class, id);
    }
}
